+++ ros_atlas/core/BasePostprocessor.py
@@ -21,7 +21,7 @@ from rospy.exceptions import ROSInitException
 
 
 class Postprocessor:
-    """Postprocessor - Base postprocessing class with abstract method for formating model-dependent message types.
+    """Postprocessor - Base postprocessing class with abstract methods for formating model-dependent message types.
     Functions
         + Handles ROS nodes related operations: Loading corresponding processor, ROS node initialization, message type formatting, faciliitate node communications, shutdown protocol
         + Subscribes to @topic:/acl_inference/<model_name>
